+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- **Logger**: Support for logger inheritance with `child` parameter
+
+### Fixed
+- **Logger**: Log level is now case insensitive via params and env var
 
 ## [1.0.2] - 2020-07-16
 ### Fixed